loader: Create framework for supporting write transactions
Implementation note: In a previous version of this feature branch, the
abstract `FileHandler.write_transaction` method used to be implemented
with `@contextlib.context_manager`. This caused MyPy to recognize it as
returning a `contextlib._GeneratorContextManager` instead of a generic
`typing.ContextManager`, which effectively locked subclasses into using
`@contextlib.context_manager` as well if they wanted to be compliant to
the static type checker. The more verbose way that is used now does not
suffer from this issue.