model: Refactor write support
The implementation of the behaviors necessary to modify a model have
been moved from the `ElementList` family of classes to the `Accessor`
family. Distinguishing between a `DecoupledElementList` and an
`ElementList` therefore no longer makes sense, which is why the former
has been renamed to simply `ElementList`. `Accessor`s that implement
write support do so with a new subclass of `ElementList`; that subclass
is dynamically created by mixing in overrides of the relevant
`MutableSequence` methods.

This new arrangement allows for more complex behaviors to be implemented
later in other `Accessor` classes, and also allows those implementations
to reuse code already in place for finding current model objects. It
also makes extending the model with new `Accessor` classes much more
straight-forward, as those will combine all the logic for finding,
adding and deleting model objects.