loader: Try to clean up on errors in `new_uuid()`
When an exception occurs during use of the `new_uuid()` context manager,
it is all too easy to accidentally leave some XML elements in the tree
that should be cleaned up. This is especially true for the "user error"
case of not calling `idcache_index()` before the `with:` block exits.
This commit introduces a small cleanup routine that attempts to fix this
error case.