gitfilehandler: Log normal stderr with INFO level
If git did not indicate an error (i.e. it returned 0), the messages it
prints to stderr are usually informative to the user and not indicative
of a problem. Therefore they now get logged with INFO level instead of
ERROR as before. For silenced git commands, stderr is kept at DEBUG.