reqif: Rework how `RelationsList`s work
Instead of assuming that the "other" objects are on the same `side` of a
relation, we now check for each relation which side we are, and return
the respective other side.

`Requirements` objects now provide a new `related` attribute similar to
the existing `relations`. In contrast, `related` works like the
`requirements` attribute on non-ReqIF objects, in that it directly
provides the respective "other" element of the relation. This makes
accessing these easier, as it's not always obvious which side of a
relation the original element is on.

Filtering the `RelationsList` for incoming, outgoing and internal
relations is now done with the `by_relation_class` method, following the
`by_*` filter methods that already exist. The method accepts the three
mentioned keywords ("incoming", "outgoing", "internal") as string
argument to filter for.