model: Beautify Jupyter notebooks
This commit provides a basic implementation for the `_repr_html_()`
methods on GenericElement and ElementList, as well as the `_repr_svg_()`
method on Diagrams. These are used by IPython/Jupyter notebooks to
provide better readable output in HTML and SVG format respectively.

The same HTML table is also provided as return value to the `__html__()`
dunder method, which is used by e.g. the `markupsafe` library.