loader: Move `entrypoint` from FileHandler to MelodyLoader
At the point in the stack where the `FileHandler` is sitting, it
shouldn't need to know what kind of relation the files handled through
it have with each other. Therefore the responsibility for determining
the model's `entrypoint` has been moved to the `MelodyLoader`. This
makes most sense, as that has always been the only class actually using
that information (apart from some logging).