model: Add external diagram caching
The external cache can be enabled by passing the new `diagram_cache` and
(optionally) `diagram_cache_subdir` to the `MelodyModel` constructor.
The mechanism internally uses the `FileHandler` infrastructure, which
means it supports loading from anywhere where loading a model itself is
possible. However, the `path` and `diagram_cache` are not required to
use the same location, or even the same `FileHandler` class. This allows
for significantly greater flexibility in the mechanism.

Currently only reading from such a cache is supported. In order for it
to be useful, it therefore has to be pre-populated with the appropriate
files. This can be done with a short script such as:

```
cache_dir = pathlib.Path("/somewhere")
for i in model.diagrams:
    cache_dir.joinpath(i.uuid + ".svg").write_text(i.render("svg"))
```