tests: Remove network-dependent model loading tests
Automated tests should not rely on the global internet being available
to them, as they might run in a restricted environment. Instead, we
should monkeypatch away the relevant functions.

Additionally, the `git+ssh://` test cannot work even with the global
internet being available to it, because it requires an SSH private key
to have been set up beforehand. That might be the case on some (or even
most) developers' machines, but most definitely not in the Github runner
or similar containerized environments. ;)