filehandler: Fix handling of single protocols
The `get_filehandler()` function unconditionally removed the first
protocol of the URL, even if there only was one to begin with. This
works fine for multi-protocol URLs (e.g. `git+ssh://...`), but breaks
for single-protocol ones (like `https://example.com`). The latter
example resulted in `//example.com` being passed to the actual
`FileHandler` instance instead of the full URL.