ci: Make pylint errors/warnings actually fail the pipeline
17f208c0c24983a304b7030ec97968ad758a5952 made the `pylint` linter run
always pass. As this was not even mentioned in the description, let
alone the reason explained, I assume it to have been unintentionally
committed along with other changes in the same file.

Note that several less important pylint warnings are still ignored in
terms of pipeline success by `|| exit $(($? & ~24))`, as well as
`-dfixme` (which would otherwise be a warning and thus fail).

Consequently, the pylint warnings that have been made visible by this
change have been fixed as well.