ci: Don't keep historic docs around
Keeping the history of the auto-generated documentation arguably has
little value. If it's needed for a specific commit, one can simply
follow the workflow steps to generate it locally. On the other hand,
with how it's maintained as a separate branch in the same Github repo,
the history clutters all commit graph views with a (usually)
uninteresting branch.