gitfilehandler: Fix erroneously requiring LFS
Even when LFS was not actually in use, the git file handler required
that it be installed. Instead of addressing the actual issue with a
simple three line patch, other contributors so far instead opted for
conditionally disabling the affected test cases. With this commit, the
git tests will run on all platforms again, and the file handler no
longer crashes when LFS is not available.

As a side effect, when LFS is indeed used (but not available), a
different error will be produced, complaining about an invalid XML file.
This is because the LFS pointer is handed back as XML without
dereferencing it. Future commits may address this matter, e.g. by
improving upon the check whether a file is an LFS file or not.