filehandler: Accept `str` as file name for `open()`
In addition to a `pathlib.PurePosixPath`, the `open()` methods of
`FileHandler` implementations now also accept plain `str` for the file
name to open. This implies that the appropriate conversion has to be
done in the file handler itself. However, as the first thing all current
implementations do is call `helpers.normalize_pure_path()` on the
argument anyway, this should not be a problem.