model: Add `GenericElement.parent`
Having access to the logical parent object from the XML point of view
has proven useful enough to add it globally.

Note that this does not supersede other `ParentAccessor` instances (e.g.
`.owner` of the `Component` types), as those are either constrained to
specific object types or mirror an attribute name used in other classes.