reqif: Split up `RequirementsAttribute`
Instead of handling the "simple" attribute types (i.e. all except
Enumeration) all in one class, we now use several different classes –
one per attribute type. This more closely matches the package structure
in Capella.

More importantly however, it allows to consolidate most of the
information on how each class works inside the class itself, instead of
spreading it out into multiple different places (a dictionary near the
top of the file, the `ValueAccessor` class which uses that mapping, and
the `RequirementsAttribute` class which in turn defers to
`ValueAccessor` for the heavy lifting). The new code can largely make do
with the simple `xmltools.AttributeProperty` primitive, eliminating the
need for the complex `ValueAccessor` class altogether.