loader: Support multiple resource locations
This commit implements support in the `MelodyLoader` for loading files
from several different locations. This functionality is needed to
support linked library projects.

Library objects can be referenced via `platform:/` links, which are now
correctly interpreted as well.

Currently only the "primary" resource will be saved when `save()` is
called on a `MelodyLoader` instance. This means that all modifications
made to elements that are part of a supplementary resource, for example
elements from a linked library, will be lost once the `MelodyLoader`
object is garbage-collected.

Fixes #66