loader: Accept `str` or `dict` as resources
This commit changes the MelodyLoader to accept simple strings or dicts
as values in its `resources=` dict. Previously, only fully constructed
`FileHandler` instances were accepted. This greatly improves usability,
as callers can now simply pass in a path or URL instead of having to
construct a FileHandler object. If that is not enough (e.g. if
credentials are needed), a simple `dict` with the named arguments can be
used now as well.