model: Register architecture layers as xtype_handlers
This is primarily useful in library situations, where it can be used to
access a library's layers.

This enables access both via `model.by_uuid()` if the UUID is known, and
via `model.search()` if not. Note that the latter *will* find the
relevant layers from the base model as well as all linked libraries.

This is a first step towards the API part of #68.