refactor(loader): Split up `get_filehandler()`
This commit introduces the new public functions `split_protocol()` and
`load_entrypoint()` in the `capellambse.loader.filehandler` namespace.
Each performs a part of what `get_filehandler()` used to do, which
enables third party code to more easily introspect the choices that
`capellambse` would make. For example, it is now possible to only call
`split_protocol()` with the `path` that `MelodyModel` would take, and
introspect which specific type of `FileHandler` would be selected for
that path.