feat(git): Complain if the user doesn't close files
If files that were written to during a transaction have not been closed
before the transaction ends, all content changes in that file will be
lost. This commit introduces logic that keeps track of open files in a
transaction and logs a CRITICAL message if any files are left unclosed
when the transaction is being finalized.