feat(git): Create a worktree per model
With this commit, the `GitFileHandler` creates and manages a git
`worktree` for each loaded model instance. When the instance is
finalized, the worktree is automatically removed again. All git commands
are executed inside the worktree instead of the bare repository clone.

Compared to working directly inside the bare clone, using a worktree has
the advantage of a separate `HEAD` reference, as well as a separate
index file (or rather, it has an index file at all). This will allow us
to simplify the write transaction code in the future. More importantly
however, this causes `git-lfs` to correctly identify which files in the
current `HEAD` use LFS and which do not. This addresses a part of #44.