fix(git): Use `git check-attr` over `git lfs ls-files`
`git lfs ls-files`, while certainly useful as a higher-level tool, has
two important drawbacks for our use case.

Firstly, it has no `-z` option and always uses `\n` as file name
separator. This is unsafe, as `\n` is a legal character for file names
within git. (In reality, this is likely a non-issue though, as LFS seems
to have far bigger problems with newlines in file names.)

The second, and far more important, drawback is that `git lfs ls-files`
is only available if `git-lfs` is actually installed. Therefore we
cannot rely on it for detecting whether or not we *need* it. Using
standard git plumbing commands instead allows us to detect this and show
a useful message to the user, instead of crashing with a very cryptic
LXML error about malformed XML.

Fixes #44