feat(model): Introduce `search(below=...)`
This feature allows to constrain a search on the model to only elements
that are nested below the given other element. The new `search()`
parameter, `below=`, accepts any semantic model element (i.e. anything
except for diagrams). The list of elements that is then returned from
`search()` is filtered for those that are nested below the given
ancestor element.

In some cases, this parent-child relationship is directly exposed as the
`parent` attribute on a class. However, note that this is very much
distinct from the `owner` attribute that is generally more common. The
latter only places a link instead of actually nesting the elements,
which is not picked up by this new functionality.

This implements "Option C" as described in #17.