refactor: Use SPDX license header in source files
This two-line license header is much less overhead compared to the
example provided with the full Apache-2.0 license, while providing
roughly the same information (considering that the full license is
provided in the repo root as well).

Additionally, a new pre-commit hook was added that automates insertion
of these headers.

Fixes #80.