refactor(aird): Fully disable isort in `aird/__init__.py`
The `isort: skip` comment does work in this specific context, but it is
very brittle for keeping an import at the top. As soon as any other
import is added before this commented line (e.g. a `__future__` import),
the manual ordering breaks.

So instead, we just tell isort to ignore this entire file. It's really
only a collection of exports anyways, so nothing of value lost here.