feat(git): Keep local caches shallow
When cloning a git repo, by default the entire history is cloned as
well. However, for the typical `capellambse` use-case of investigating a
model at a specific point in time, the history is irrelevant. Fetching
it anyways means wasting network bandwidth and disk space for data that
is probably never used.

On one of our production repositories, this reduces the amount of data
transferred by about 90%, and the time until the `MelodyModel` object
becomes ready during an initial clone by roughly the same amount.