refactor: Implement `_repr_mimebundle_` on model Diagrams
The unique situation with the diagram cache requires a bit more delicate
control over the MIME types handed to the front-end for displaying a
diagram than is possible with dedicated methods per MIME type. The MIME
bundle approach allows us to exert that control. This commit fixes a
concrete issue with cache handling, which was the original reason for
implementing this change.

It is possible to have a cache for one image format, but not others
(e.g. cached SVGs, but not PNGs). In this case, for the non-cached image
formats, the diagram would be rendered by the `capellambse.aird` module.
This could lead to differences in the displayed diagram depending on
which format the frontend selected.

Instead, if a cache is available and at least one of the allowed formats
in the MIME bundle is cached, then only cached data will be returned. If
no format can be loaded from cache, or no cache is available in the
first place, the supported formats will be rendered by `capellambse` as
usual.