Solves #108: Fix default attribute setting
Default values are not written to the xml by Capella. We don't do this
too. This commit accomplishes the following:
- Fix order of parameters such that "id needed" exception isn't raised.
- Be safe when deleting default values on `obj`. We changed L255 and
L309, not removed try except block in `AttributeProperty` `__del__`. The
exception should be raised there. It makes sense to call `__del__` from
`__set__` iff the attribute exists in the xml.
- Instead of just testing `BooleanAttributeValue` we test
creation with default value on all possible AttributeValues.