fix: Only allow PNG in untrusted notebooks
The `jupyter_untrusted` setting was introduced to work around issues
with Github's notebook preview renderer. As it turned out, these issues
not only occur with SVGs that use `<style>` sheets, but also `<img>`
whose `src` is a `data:` URI - the latter has become exposed to the
frontend with the previous commit.

To fix this new issue and prevent future ones similar to it, we now only
offer the PNG format to the frontend, instead of offering everything but
SVG.