fix: Improve handling of HTML attributes in the model
Attributes that use HTML instead of plain unformatted strings had caused
some surprising problems. Most notably, the "round-trip check" when
setting them triggered an exception when they were fed HTML that wasn't
escaped like `markupsafe` would. This did not only apply to malformed
HTML, but also valid HTML that was just using fewer XML entities.

This was caused by the interplay between the `AttributeProperty`'s
round-trip check and the simplified way that we used of telling it to
return `markupsafe.Markup`. This led to the cleaned-up and properly
escaped HTML being compared to the raw input. As one would expect, this
had a high chance of failing the round-trip check.

Therefore the new `HTMLAttributeProperty` class was introduced, which
cleans and escapes the given input *before* handing it over to the base
`AttributeProperty` and its safety checks.