refactor(model): Split up the `ProxyAccessor`
The `ProxyAccessor` has been split up into multiple different classes,
each taking over a part of the former `ProxyAccessor`'s
responsibilities:

- The `DirectProxyAccessor`, which handles the most common simple cases
- The `ReferencingProxyAccessor`, which handles the case where the child
  elements in the tree are only references onto the actual elements, and
  not the elements themselves (this takes away the `follow=` argument)
- The `DeepProxyAccessor`, which searches through an entire subtree
  (this takes away the `deep=` argument)

This split simplifies the logic around the various `__get__` methods of
these accessors. It also allows to implement the different strategies
for the related `__set__` methods in a cleaner way in a future commit.

Be aware that the default value for `follow_abstract=` to
`ProxyAccessor` has been changed from `True` to `False`, along with a
slight, but significant, semantic change. Previously it would *try* to
follow an `abstractType` reference if it was `True`, and would fall back
to yielding the element itself if `abstractType` was not set. This has
been changed: It is now an error if `follow_abstract` is `True`, but
`abstractType` is not set in the XML.

All references to the `ProxyAccessor` have been updated.

Co-authored-by: Martin Lehmann <martin.lehmann@deutschebahn.com>