feat(accessors): Allow create/delete of elements with rootelements
Before only create of elements behind a `WriteableAccessor` that had
elmlist declared was enabled. Calling create on accessor-attributes that
were declared with the rootelem parameter would just raise a TypeError.
This commit unlocks this functionality and hints the requester on
passing a root object in `kw`. The `etree._Element` of the given root
can now be taken as the parent which makes the mandatory `elmlist`
parameter now optional. The `__get__` of the `DirectProxyAccessor`
returns itself if `_make_list` returns `None` such that create can be
called.

In order to delete elements that are accessible from non `ElementList`s
a delete method was added to `common.GenericElement`.