feat: Add a simple REPL launcher
This script launches a simple Read-Execute-Print loop based on the
stdlib `code` module. It is intended to facilitate quick and simple
debugging action where it would be too much overhead to launch a full
blown Jupyter server. However, it is also very useful as a quick and
easy way to look up some things in a model.

For convenience, VS Code compatible launch configurations were added for
the models maintained in the `capellambse` source tree.