feat(loader): Check for duplicated UUIDs
This commit introduces new model validity checks in the `MelodyLoader`
and `ModelFile` classes. While loading a model, it is now checked for
UUID duplication both within as well as across fragments. Before saving
a model, a check for duplication across fragments is performed as well.
The check can also be triggered manually by calling the new
`check_duplicate_uuids` method on the `MelodyLoader`.