fix(loader): Use `set`s for the `xsi:type` caches
This avoids an issue where elements would show up in the xtype-cache
multiple times if the `idcache_index()` method was called multiple times
on the same subtree.

This commit also slightly improves the performance when deleting
elements, as finding the right element in the cache is now O(1) instead
of O(n).