revert(model): Remove dict-based element creation logic
This commit partially reverts dd8e7d5411b6c48ea735c2962cd1345124f6b734.

Unlike the new design, the "tedious and old-fashioned API" was stable
and (at least somewhat) reliable, and the dev team agreed that the
functionality of creating nested complex objects should rather be
implemented in a separate module.

The logic that allowed to `__set__` certain attributes to a list of
strings was also part of the aforementioned commit. Instead of reverting
that as well, it was moved to the `Accessor` classes.