docs(model): Fix false-positive pydocstyle lint error
pydocstyle recognized the "Examples:" mini-heading before a list of
examples as a proper heading, and complained that it wasn't styled like
one. This also subsequently led to more complaints, because it didn't
recognize the parameter documentation following this line as such. This
commit changes the mini-heading to "Some examples:", which avoids the
confusion.