fix(model)!: Reimplement the `ReferencingProxyAccessor`
This commit replaces the `ReferencingProxyAccessor` with the new class
`LinkAccessor`. Instead of inheriting from the `DirectProxyAccessor`
class, this new accessor directly inherits from `Writable`- and
`PhysicalAccessor`.

The behavior towards a user of `capellambse.MelodyModel` did not change,
however the implementation was rewritten from scratch, and the c'tor
arguments changed. The new implementation is aware of the XML tags in
use, which enables it to create new tags and therefore add new links to
a given list. However this only applies to client classes which pass in
tag information – this is currently only the case for the unit test
subject, `LogicalComponent.allocated_functions`. Other classes currently
pass `None` as the tag, which raises a `DeprecationWarning` and should
be disallowed again once all classes have been extended to pass in the
correct tag.

BREAKING CHANGES: The (broken) `ReferencingProxyAccessor` was removed;
use the new `LinkAccessor` instead. Extensions that used it need to be
updated.