feat(decl): Implement forward-references via promises
Objects in the Declarative Modelling YAML can now be assigned a
`promise_id`, which is a unique, user-defined string. These IDs can
later be referenced with the `!promise` YAML tag.

If the same `promise_id` is assigned to multiple objects, an exception
is raised informing about the non-unique ID.

If a `!promise` tag is encountered that is never fulfilled while
processing a YAML file, an `UnfulfilledPromisesError` is raised. This
exception's first (and currently only) argument is a `frozenset`
containing all unfulfilled `promise_id`s.