fix(model): Don't invent names for unnamed elements
Elements that did not have an explicit `name` set on them made up a name
string indicating as much. This made it unnecessarily hard to
programmatically distinguish when a name was explicitly set.

With this commit, the `name` attribute will simply be empty if there was
no name set on the element.