fix(model): Remove legacy `__str__` from ElementList
This method was intended to produce easier legible output than
`__repr__`, however now that repr is human-readable (and even more
useful than str was), this is no longer necessary. Removing the
`__str__` method makes e.g. `print(elmlist)` behave just like the Python
REPL when it's printing an attribute that is a list.