feat(model): Implement the terminal graphics protocol
This format generates escape codes containing the rendered PNG image
according to the [terminal graphics protocol]. These escape codes are
also included in the return value of `AbstractDiagram.__repr__` if it
detected that a supported terminal is connected to the three standard
streams.

This initial implementation for the detection logic only looks at the
`$TERM` environment variable. Future commits may extend this by actively
querying the terminal in order to improve detection rates.

[terminal graphics protocol]:
  https://sw.kovidgoyal.net/kitty/graphics-protocol/