refactor(model): Stop using `CustomAccessor`
It's been deprecated for a while now, so migrate away to appropriate
alternatives (which, in all these cases, is the
ReferenceSearchingAccessor).

The ref searcher has also gained the ability to work with multiple
target classes, which can be passed in using a tuple of classes in place
of a single class.