fix(model): Ignore `@xtype_handler` in `__main__` module
When a single module simultaneously provides new `@xtype_handler`
classes and a script entrypoint, the module is executed twice: The first
time, `__name__` inside the module is set to `__main__`, and the second
time – triggered by `capellambse.load_model_extensions()` – it is set to
the module's actual name. This causes all classes in it to be registered
twice, which is rejected with an exception. This also causes the
initialization function of the extension to not be exceuted, which would
result in broken functionality.