refactor(diagram_cache): Extract common "native Capella" logic
This commit refactors the `update_diagram_cache` method and the module
behind it. A new submodule `capellambse._native` bundles logic to
facilitate working with a native Capella installation, both as plain
executable and as Docker image. The diagram cache related functionality
was greatly simplified thanks to the extracted native-Capella logic.