fix(filehandler): Don't crash with local files if git isn't installed
If there's no `git` executable to begin with, the exception flying isn't
CalledProcessError but FileNotFoundError, which was previously not
handled at all.

Since the data provided by the affected `__git_rev_parse` function is
optional to begin with, this commit changes it to suppress (and log)
every Exception, not just those of a specific subclass.