feat(glart): Look harder for a recent successful job
It turned out that the former default search size of "the 200 most
recent successful jobs" was too conservative in practice. This number
was increased to search up to 1000 jobs by default (thus making up to 10
API requests). This limit can now also be changed with the environment
variable `CAPELLAMBSE_GLART_MAX_JOBS`.