fix(ci-templates): Only push derived branches from the default branch
Previously, pushes to other branches would result in the derivation job
running (as it should), but also lead to the "derived/*" branches being
pushed. This was a very unsafe default, as it could easily lead to
production results being overwritten with test data.

This commit changes the default job setup so that the "derived/*"
branches are only pushed from the project's default branch (usually
main). Other branches still produce the artifacts archive, but do not
cause the result to be pushed as branches.