ci: Rework Code Quality related workflows
This commit combines the build/test/publish and lint workflow files into
one, as they always have the same triggers and are very closely related.

Additionally, the combined workflow is adjusted so that it properly
triggers on opened pull requests. Note that it no longer triggers on
pushes without an associated PR (except for master itself), but if need
be, those workflows can be manually run via workflow_dispatch.