fix(diargam_cache): Add `_sanitize_svg` with new overrides
Somehow Capella's SVG writer sets `fill` and `stroke` to `black` on the
root SVG element. These styles get inherited from `clipPath` elements
which might span the whole SVG viewbox. In a PDF the inheritance is
considered such that instead of a diagram, only a black box is rendered.

This commit fixes the issue.