feat(model): Disable AIRD renderer if diagram_cache was given
Experience has shown that it's more often than not undesirable to have
the internal rendering engine jump in for diagrams that could not be
found in the cache. This commit arranges for the internal engine to be
disabled for AIRD diagrams by default if a cache was specified. This
behavior can be overridden by setting the new ``fallback_render_aird``
model load parameter to ``True``, which re-enables the rendering engine
for all diagram types.

Diagrams that do not come from one of the model's AIRD files are not
affected, and will always fall back to the internal engine. This means
that e.g. context-diagrams, which are generated on-the-fly, will
continue to work as expected, without the need to pre-generate them.

Resolves #287.