feat(diagram_cache): Improve generated index files
This commit makes some improvements to the diagram cache index files
that can be optionally generated.

The JSON file now contains the new "viewpoint" and "type" keys for each
diagram, which specify the viewpoint (i.e. architectural layer) that the
diagram belongs to and the diagram type.

The HTML index has gotten a somewhat larger overhaul. The stylesheet has
been updated to provide a sleaker, more modern look, and make the UUIDs
more easily visible (but still not overbearing). The diagram names are
now grouped by the viewpoint they belong to, and each diagram also has a
prefix containing the short type code. When hovering over the type code,
the full name of that diagram type is displayed.