fix(model): Skip displaying ReferenceSearchingAccessor attributes
ReferenceSearchingAccessor becomes very slow to compute very quickly as
a model grows. When displaying an object, this can cause a delay of
several seconds up to multiple minutes, during which the prompt or
notebook appears stuck and unresponsive.

This commit omits all attributes using ReferenceSearchingAccessor from
the display routines used by the Python REPL and Jupyter notebooks.
Instead, only a short text informing about the nature of the attribute
is shown.

Resolves #295.