ci: Update pre-commit hooks
Update all hooks with `pre-commit autoupdate`.

Most notable is the updated `docformatter` version, which now has an
opinion on the blank line between the file copyright comment and the
module docstring (it says there shouldn't be one).

The only other change is that the hooks have been reordered in the
pre-commit config file (yes, once again :) ). Now the general ordering
is formatters first, then linters, because formatters are generally much
faster than linters and it might be necessary to touch the formatted
files manually again. This is much more convenient if we don't have to
wait for the linters to see all formatters' outputs.