feat(cli_helpers): Add helpers for loading info about a model
Previously there were only functions that would already load the model.
However, this made it impossible to change anything about how the model
is loaded, which can be handy e.g. when loading a "known model".

This commit fixes this oversight by adding CLI helpers which load the
model description dictionary instead. It also adjusts the REPL to use
the new helpers instead of rolling its own code for this feature.