feat(aird): Add a faster way to get visible diagram elements
This commit adds the `iter_visible` function the the AIRD parser module,
which yields all semantic elements visible in a given diagram. Unlike
`parse_diagram`, it doesn't build the whole diagram and all elements,
which saves a lot of time.

This new function is then used to significantly accelerate access of
`Diagram.nodes` in the `model` package.