fix(filehandler): Move lockfile into $GITDIR
For local git repos, the locking mechanism would put the
`capellambse.lock` file into the root of the worktree. This was
annoying, as that file had to be explicitly gitignore'd or it would show
up in diffs and `git status`.

This commit changes `self.__repo` to point to the $GITDIR instead of
the main worktree root, thereby moving the lockfile there. This more
closely mirrors the behavior of remote git repos, which don't have a
main worktree and therefore also have `__repo` point to the $GITDIR.