fix(model): Fix `AbstractDiagram.nodes`
The `nodes` property of `AbstractDiagram` used to assume that we were
talking AIRD diagrams. Not only was this assumption misplaces, it was
also made obsolete by 04c695a18a840bbd350458646cb68fcf4969998c, which
added an override on the specific subclass.

This commit fixes the default implementation to make the (more
reasonable) assumption that the IDs of diagram elements are those of the
semantic elements.

Co-authored-by: Martin Lehmann <martin.lehmann@deutschebahn.com>