feat(loader): Derive entrypoint from filehandler's listing
This commit makes the MelodyLoader use the new FileHandler capability to
enumerate files in order to automatically derive the `entrypoint` in
more cases. It will now look into the filehandler's root directory, and
if there is exactly one `.aird` file, that becomes the new entrypoint.

If there is more than one `.aird` file in the root directory, an error
will be raised, and the same is true if the root directory does not
directly contain any `.aird` files. In these cases, it is still
necessary to manually specify the `entrypoint` at load time.