feat(native): Copy files into / out of docker container
The Docker runner for capellambse's Native Capella functions used to use
volume mounts for making the project and workspace available to the
container. While it's the simplest approach to implement, it has a few
drawbacks. The most notable one is that it doesn't work properly if
capellambse itself runs within a container and was given permission to
use Docker-in-Docker with the host's docker engine. As CI (capellambse's
primary target environment) is usually based around docker containers or
something similar, this drawback has become unacceptable.

This commit therefore changes the behavior of the Docker runner to copy
the project and workspace files into the spawned "slave" container
before executing Capella inside of it, and to copy the modified files
back to their original locations afterwards.