perf(git): Prefetch is_dir/is_file results
This commit changes `GitFileHandler.iterdir()` and `_GitPath` to consume
type information directly from the `ls-tree` call instead of querying it
on request with an extra call to `cat-file -t`.

This change drastically improves the performance of `walk()`, as it now
only requires one subprocess call per tree, instead of two per tree and
one per blob.