feat(cli_helpers): Log when starting to load a model
This log message makes it slightly more transparent what is currently
going on during startup. It also makes it easier to correlate load
errors with the model that caused them, in case multiple models are
involved. This change is most notable with the Click parameter types, as
loading of these usually happens before other log messages from the
application can be output. Unfortunately however, this is also where
it's the most difficult to show its effects, as the logging setup also
often is done as part of the application code and therefore after the
models have already been loaded. Despite that, it still makes sense to
have this message for all cases where logging is initialized before
parsing command line arguments.