fix(model): Display name of element in `__html__` heading
In the `__html__` representation, which is used for Jupyter, the name of
the type of the element was displayed in the heading at the top.

While the type is relatively uninteresting for the user, especially
because the xtype is displayed right after it, the user-specified name
of the element was only shown as part of the full list of attributes.

This commit replaces the type name in the heading with the user-given
name if available.