fix(loader): Pass `**kwargs` on to `get_filehandler`
With the new entrypoint logic introduced in
edd2c3c83bf8b66547099872056f1b07e9ba7091, additional keyword-only
arguments that needed to be passed to the `get_filehandler` function
would get lost. This could cause inexplicable follow-up errors like
"permission denied" or "file not found", when it dropped login
credentials or a VCS revision argument.

This commit also changes the
`ignore_duplicate_uuids_and_void_all_warranties` argument to be a
"proper" keyword-only argument to MelodyLoader, instead of pulling it
from `kwargs`. This caused pylint to recognize the `kwargs` as being
used, and therefore it did not issue an "unused" warning.