fix(filehandler): Ignore Git `credential.helper` for passed credentials
On many systems, such as MacOS and Windows, a default Git credential helper
is defined. The credential helper is set in the Git configuration
with the credential.helper option.

According to the official Git documentation (https://git-scm.com/docs/gitcredentials),
the `GIT_ASKPASS` environment variable is ignored when a credential helper is configured.

Therefore, the username and password arguments were ignored when a
credential.helper was set. This was not intentional, as the user
obviously wants to use the credentials passed.